Create initialFile
Beginning of the program. So far just designed to automate the dealer's hand, and the player simply hits until bust. Next step is adding the necessary steps to update the tables based on the results and creating a cycle through which it can learn over several thousand hands. Lot of print statements to remove in that case...