Update initialFile
Added updates to the tables, a way to play a given number of reps in a row, and a refresh to create a new set of data without quitting the program. The learning is not quite where it should be, winning about 40% of the time in a sample million. There is still not handling for aces, either. That'll be the next step, then refinement of the learning.